ignore .gitignore when ln -s

+++ dotfilesLink.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 for df in \.*; do
   # 通常のファイルのみに対してシンボリックリンクを作成する
-  if [ -f $df ]; then
+  if [ -f $df ] && [ $df != ".gitignore" ]; then
     ln -sf ~/dotfiles/$df ~/$df
     echo "ln -sf ~/dotfiles/$df ~/$df"
   fi
